Move domain records to zone records.
